third_party/blink/web_tests/wpt_internal/fenced_frame/badging.https.html

<!DOCTYPE html>
<title>Test Notification</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
promise_test(async () => {
  const frame = attachFencedFrameContext();
  try {
    await frame.execute(() => {
      return navigator.setAppBadge(1);
    });
    assert_unreached('setAppBadge executed without error; want error');
  } catch(e) {
    assert_equals(e.message,
                  "The badge API is not allowed in this context",
                  "The Badge API should cause exception in a fencedfarme");
  }
}, 'setAppBadge should fail inside a fenced frame');

promise_test(async () => {
  const frame = attachFencedFrameContext();
  try {
    await frame.execute(() => {
      return navigator.clearAppBadge();
    });
    assert_unreached('clearAppBadge executed without error; want error');
  } catch(e) {
    assert_equals(e.message,
                  "The badge API is not allowed in this context",
                  "The Badge API should cause exception in a fencedfarme");
  }
}, 'clearAppBadge should fail inside a fenced frame');
</script>
</body>
